makemigrations를 하지 않았으므로 api로 모델 이동, app 삭제
