--- movies/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-
-from .models import Movie, Review
-
-admin.site.register(Movie)
-admin.site.register(Review)
